requirements.txt
duckduckgo-search
python-telegram-bot==20.7
requests
yfinance